requirements.txt
Django==4.2.24
djangorestframework==3.14.0
dj-rest-auth==5.0.2
django-allauth==0.57.2
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.3.1
django-environ==0.11.2
psycopg2-binary==2.9.9
